Round v5 B9: prefer sinf/cosf/fabsf in renderer hot paths
clang-tidy performance-type-promotion-in-math-fn flagged 51 sites
where float-typed inputs widen to double across math.h calls. PPC
double-precision math is ~2x scalar single-precision (G3) and pulls
work off the AltiVec single-precision pipelines (G4). Even on Intel
where the difference is small, fewer cycles is fewer cycles.

This commit converts the per-frame-hot ones (rest are cold paths
or QC builtins, separate batch):

* mathlib.c:177-184 -- AngleVectors does 6 sin+cos per call, called
  per visible entity per frame. Hottest math site in the engine.
  sin->sinf, cos->cosf.
* r_brush.c:1019 -- R_AddDynamicLights does fabs(dist) per dlight
  per surface. fabs on a float widens to double, function call;
  fabsf is a single-instruction sign-bit clear on PPC.
* r_alias.c:854-855 -- alias shadevector cos/sin; per alias entity
  with shadows. Also fixed the surrounding constants from double
  literals to float literals (16.0f, 2.0f, 3.14159f) so the
  multiply chain stays in float.
* gl_rlight.c:112-113 -- R_RenderDlight billboard ring (only used
  with gl_flashblend; minor but cheap fix).

Numerical equivalence: float math has fewer mantissa bits but at
game-time precision (positions in 0..8192 unit world, angles in
0..360 deg) the difference is well below 1 LSB of the rendered
output. Visually identical.

Cycle savings rough estimate (G3, AngleVectors per call):
  before: 6 calls to libm sin/cos via float->double->float =
    ~6 * 30 cyc + 12 conversions = ~210 cyc
  after:  6 calls to sinf/cosf =
    ~6 * 20 cyc = ~120 cyc
  ~40% reduction on the math-heavy path.

+++ Quake/gl_rlight.c
@@ -107,10 +107,10 @@ void R_RenderDlight (dlight_t *light)
 	glColor3f (0,0,0);
 	for (i=16 ; i>=0 ; i--)
 	{
-		a = i/16.0 * M_PI*2;
+		a = i/16.0f * (float)M_PI*2.0f;
 		for (j=0 ; j<3 ; j++)
-			v[j] = light->origin[j] + vright[j]*cos(a)*rad
-				+ vup[j]*sin(a)*rad;
+			v[j] = light->origin[j] + vright[j]*cosf(a)*rad
+				+ vup[j]*sinf(a)*rad;
 		glVertex3fv (v);
 	}
 	glEnd ();
